process dual and merge references: https://github.com/metanorma/metanorma-ogc/issues/184
